Re #952 Upgrade Grails to 2.5.4, get all tests to pass

+++ grails-app/domain/us/wearecurio/model/AnalyticsCorrelation.groovy
@@ -72,7 +72,7 @@ class AnalyticsCorrelation {
 		series2Type index: 'updateIdx'
 		description1 index: 'search1'
 		description2 index: 'search2'
-		signalLevel index: 'signalLevelIdx', defaultValue: (Double)-1.0
+		signalLevel index: 'signalLevelIdx', defaultValue: -1.0d
 	}
 
 	public static userCorrelations(Long userId, Integer max, String flavor) {
